fix(Accounting):fix foreign location_id in accounting_mapping_setting…

+++ Modules/Accounting/Database/Migrations/2024_01_25_144122_add_cloumn_business_locations_to_table_accounting_mapping_auto_migrations.php
@@ -14,10 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('accounting_mapping_setting_auto_migrations', function (Blueprint $table) {
-            $table->bigInteger('location_id')->unsigned()->after('id');
+            $table->integer('location_id')->unsigned()->after('id');
             $table->foreign('location_id')->references('id')->on('business_locations');
 
-           
         });
     }
